feat : method query params to paginate

+++ dto/PaginateDto.go
@@ -2,5 +2,5 @@ package dto
 
 type PaginateDto[T any] struct {
 	Data        T   `json:"data,omitempty"`
-	TotalRecord int `json:"totalRecord"`
+	TotalRecord int64 `json:"totalRecord"`
 }
